+++ src/index.tsx
@@ -7,3 +7,4 @@ export {
   VariableContextProvider,
 } from "react-native-css";
 export * from "./compat";
+export * from "./metro";